+++ src/main/java/com/simibubi/create/compat/emi/CreateSlotWidget.java
@@ -1,0 +1,34 @@
+package com.simibubi.create.compat.emi;
+
+import java.util.List;
+
+import com.simibubi.create.foundation.utility.Lang;
+import com.simibubi.create.infrastructure.config.AllConfigs;
+
+import com.tterrag.registrate.util.entry.FluidEntry;
+import dev.emi.emi.EmiUtil;
+import dev.emi.emi.api.stack.EmiIngredient;
+import dev.emi.emi.api.stack.EmiStack;
+import dev.emi.emi.api.stack.FluidEmiStack;
+import dev.emi.emi.api.widget.SlotWidget;
+import dev.emi.emi.config.FluidUnit;
+import net.minecraft.ChatFormatting;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
+import net.minecraft.core.Registry;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.util.FormattedCharSequence;
+import net.minecraft.world.level.material.Fluid;
+import net.minecraftforge.registries.ForgeRegistries;
+
+public class CreateSlotWidget extends SlotWidget {
+	public CreateSlotWidget(EmiIngredient stack, int x, int y) {
+		super(stack, x, y);
+	}
+
+	@Override
+	public List<ClientTooltipComponent> getTooltip(int mouseX, int mouseY) {
+		List<ClientTooltipComponent> tooltip = super.getTooltip(mouseX, mouseY);
+		return tooltip;
+	}
+}